+++ global_var.py
@@ -13,3 +13,6 @@ trace_list = []
 numerical_error_tol_abs = 1e-12
 numerical_tol_bit = 12
 numerical_error_tol_rel = 0.01
+
+flop_error_tol_abs = 5e-7
+flop_error_tol_bit = 7