+++ run/one_click.sh
@@ -1,0 +1,3 @@
+sh run/compare_glb.sh
+sh run/static_optm_test.sh
+sh run/compare_dyn.sh